(QENG-1640) Use more explicit path for fact count
Previous to this commit, the location of a text file containing how many
facts to use per agent was dropped in /etc/puppetlabs/num_facts. However
that is not very descriptive and does not let the user know what that
file is there for.

This commit changes it to look for the file now in
/etc/puppetlabs/clamps/num_facts instead.