(maint) Fix amqserver to use $::settings::server
Previous to this commit, the code was accidently changed from using
$::server as the default. However that returns undef as it does not
exist as a fact. Its current default was set to $::servername, however
that is the same as $::fqdn, which in ec2 will not match the certname.
This commit changes it to use $::settings::server, which will pull the
value out of puppet.conf and use that value.