Remove break
Prior to this commit I had a `break` in here that actually should have
been a `next`. However on review I actually did not want this at all
as we want an empty hash and this was in here only before I had figured
out that evaluate returned native ruby objects. This commit removes the
line completely