(QENG-2261) Add and use `$splay` and `$splaylimit` options
For scale testing we wish to support enabling `splay` on non-daemonized agent
runs. We will pass this in from the config file. Here we pass it down through
the classes to ultimately set the `--splay` and `--splaylimit <limit>` args
to the puppet agent.

A related pull on https://github.com/puppetlabs/pe_acceptance_tests will be
necessary.