(QENG-2464) Add support for static files
Prior to this, CLAMPS generated only constantly changing files.
This method of generating files ends up decreasing the resource
duplication rate to an unrealistic degree, causing issues for
load testing, and having negative effects on caching, etc.

Here we're adding the ability to specify a number of static File
resources to be created. Later we intend to streamline the dynamic
File resources so they can be specified more predictably.