Fix uneven distribution of cron jobs based on fqdn_rand
Prior to this commit, the cron jobs created for each user were
based on fqdn_rand which does not gaurentee an even distribution
of agent runs on the clamps node.

After this commit, the cron jobs are created one for in each minute.
so user1 will apply at 1 and 31
user5 will apply at 5, 35
user128 will apply at 8, 38 etc...

If you use a multiple of 30 for your non_rootusers param then you
will get a perfectly event distribution of agent runs.