Allow not managing non-root mcollective settings
Prior to this commit, we always managed non-root mcollective even
if we did not want to run the service.  Managing the settings
has an issue where they change every run when you point the clamps
agent at a load balancer.

After this commit, we will not manage mcollective for non-root users
when we're not running the mco service for them.  This allows an
easy workaround for pointing the clamps agent at a load balancer.