(ORCH-1652) Use mock pxp-module-puppet for cron job
When using a mock pxp-module-puppet and cached catalogs, there's no way
to update the catalog used, so real agent runs are largely irrelevant.
They also use a lot more memory.

When using a mock pxp-module-puppet, use it for the cron job as well as
pxp-agent runs to minimize memory use.

When cached catalogs are enabled, the mock pxp-module-puppet will use
custom caching for the catalog to avoid making node and catalog requests
that would differ from a normal Puppet run.