(ORCH-1652) Use a single file for facts
Clamps agents should use custom facts, which means using Puppet to
evaluate the facts. As we don't want to do this for every clamps agent -
as it negates much of the benefits of a mock agent - instead save the
root agent's facts to a file that each agent can access.